requirements.txt
streamlit==1.35.0
openai-whisper==20250625
transformers==4.41.2
torch==2.2.2
torchaudio==2.2.2
moviepy==1.0.3
imageio-ffmpeg==0.6.0
numpy==1.24.4